requirements.txt
sentence-transformers
faiss-cpu
requests
gradio
